marketplace-scraper/requirements.txt
playwright>=1.43.0
playwright-stealth>=1.0.6
httpx>=0.27.0
beautifulsoup4>=4.12.3
google-genai>=0.1.0
gspread>=6.1.2
google-auth>=2.29.0
openpyxl>=3.1.2
pyyaml>=6.0.1
customtkinter>=5.2.2
